personal-portfolio/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.2
pymongo==4.6.0
motor==3.3.1
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.14.0
torch==2.1.1
transformers==4.35.2
sentence-transformers==2.2.2
opencv-python==4.8.1.78
pillow==10.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
celery==5.3.4
redis==5.0.1
httpx==0.25.2
python-slugify==8.0.1
python-magic==0.4.27
boto3==1.29.3
elasticsearch[async]==8.11.0
langchain==0.0.339
openai==1.3.5
